Refactor ConstantTicksToLiveExpirationPolicy
... by moving it to ExpiringSet
